backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
sse-starlette==1.8.2
httpx==0.27.0
sqlmodel==0.0.16
aiosqlite==0.19.0
